refactoring code to create submission service
